splitting up files, fixing tests
