+++ src/types/tape/familiar/menagerie/mod.rs
@@ -1,0 +1,7 @@
+// two things here,
+//
+// 1. implementations of specific familiars over specific types
+// 2. a Menagerie class which manages creating, moving/adjusting, and sharing familiars (eventually)
+//
+
+pub mod tape_familiar;
